Refined implementations of the keyboards. We can now detect and use more keys including enter, shift, and backspace.
Finally in the position to setup the data that I'll receive from the
leap in the keyboard class. I just need to get the appropriate data.
Send it over, and then start rendering or firing key events based off of
the position data.

Need to set up plane calibration as well.

After keyboards set up, I can focus on setting up the experiement
contorller and and start looking into adding the screen transition logic
for the keyboard/mouse and the leap keyboards (waiting on
controller-based and tablet-based keyboards).