TODO: In Exit Survey: Limit age to numbers only. Limit Ranking to 1 - 5 and must be unique
Fixed a small issue with the Swipe Keyboard. We now detect pressed
points as we enter the threshold of the expected key rather than at the
center. This helps us reduce false positives.

Limited the backspace so that it no longer deletes correctly typed
letters.

No longer need to run the tutorial to run practice.

Finished the Exit Survey file writing process.