Finished adding the word pattern recognizer. It can now build dictionaries based off of similar words.
I also added a "terms-to-block" dictionary in order to remove swear
words or racially sensative words. I need to make sure to have this
happen for practice words too.

The next thing to do is to change the way the experiment works so that
it will use the new dictionaries!

To do this, we need to just load the static dictionaries based on the
keyboard's file name rather than using resevoir sampling. For practicing
we should use resevoir sampling such that we don't use any of the words
that will be used in the experiment.