Currently working out a bug that could cause errors in Dictionary Building. I'm finding a way to calculate the number of words in each container based off of the word pool size and the number of containers that we calculate.
Had to work around Java's faulty precision.
In Java 1/3f != (1 - 2/3f).

To fix this I use my own precision class to round the values. The other
solution is using BigDecimal which is much more complicated and
convoluted with it's lengthy function calls. And you have to treat it
special to get around numbers such as 1/3. It doesn't handle repeating
decimals well.