Made various changes to the keyboard layouts. I reduced the layouts to follow the removal of all special keys. Backspace is now appropriately positioned on keyboards to make it easier to use as well.
Only a few more things to implement. The moving leap plane with a touch
threshold and normal that follows the normal of the tool.

And fixing the pinch threshold to work better. Might not use the
stabalized palm position because it's really hard to use on the swipe
keyboard. The normal palm position can work if we somehow correct it.

See TODO_LIST for more info.