Exit survey now split up. Hand alternative added.
Exit survey is now split up into a final exit survey and intermittent
keyboard survey. Participants can now retake a survey, the survey's will
automatically fill in any previous information so that changes are easy
without refilling out the entire survey.

The hand can now be used for any of the leap keyboards. So this needs to
be determined if it is the de-facto interaction that replaces the
stylus.