New Dictionaries, removed ID's, added DISABLED type, removed controller.
Built new dictionaries for 15 trials. Removed some of the less common
words in favor of more modern/common words so that subject have a higher
likelihood of knowing the word.

Keyboards no longer have numerical ID's and instead completely identify
by type. With this change, we can now add or remove keyboards without
breaking anything.

Added DISABLED type to the KeyboardType enum. With this, we can now
assign keyboards we aren't using as DISABLED and they will no longer be
built or stored in memory. They will also be removed from lists and
ignored in data collection and calculation.

Lastly, the controller keyboard was changed to DISABLED and will no
longer be used in the experiment. All IRB documents have been updated
and resubmitted to reflect this change.